Separate file for JsonParser and test description
