+++ task-queue/lib/parserJson.test.js
@@ -1,16 +1,22 @@
 const {
-  JsonParser,
   ParserError
 } = require('./parser');
+const {
+  JsonParser
+} = require('./parserJson');
 
+// Assert `prepare` and `check` methods work correctly
 describe('JsonParser', () => {
+  // Test object
   const obj = {
     type: 'test',
     args: [ 'test' ],
     flag: true,
     num: 32197361
   };
+  // Corrupted string
   const corruptStr = '{"type": "test" ,}';
+  // Parser
   let parser = JsonParser;
 
   it('should stringify an object', () => {
